Create bgbXML_to_JSON.py
This converts all xml files in the bgb folder from dropbox into one large json dictionary called 'bgb_json'. It is one object called "bgb_voyages", for the eventuality that we combine them all into one json, then every set of voyage sources can be an object. within the object bgb_voyages are objects for every voyage, containing key-value pairs for every piece of information. For the cargo, the values are arrays, where the index of every array corresponds to the others, ie: Quantities[1] corresponds to Products[1] and Specification[1] etc.